moved flask monitoring to its own module
